+++ packages/core/src/utils/migration.ts
@@ -2,7 +2,7 @@ import { AskarModule } from '@credo-ts/askar'
 import { Agent, ConsoleLogger, LogLevel } from '@credo-ts/core'
 import { IndySdkToAskarMigrationUpdater } from '@credo-ts/indy-sdk-to-askar-migration'
 import { agentDependencies } from '@credo-ts/react-native'
-import { ariesAskar } from '@hyperledger/aries-askar-react-native'
+import { askar } from '@openwallet-foundation/askar-react-native'
 import { Platform } from 'react-native'
 import * as RNFS from 'react-native-fs'
 
@@ -25,7 +25,7 @@ export const migrateToAskar = async (walletId: string, key: string, agent?: Agen
       },
       modules: {
         askar: new AskarModule({
-          ariesAskar,
+          ariesAskar: askar,
         }),
       },
       dependencies: agentDependencies,